# api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "私有部署企业协作产品的服务端 API。当前阶段包含管理员登录、管理员创建普通用户、普通用户登录。",
        "title": "AI 原生企业协作服务 API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/apps": {
            "get": {
                "description": "管理员读取应用配置，包含回调地址和回调密钥。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "列出应用",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.listAdminAppsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "description": "管理员创建一个应用配置。回调密钥由服务端生成。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "创建应用",
                "parameters": [
                    {
                        "description": "应用配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.adminAppRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.adminAppEnvelope"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/admin/apps/{id}": {
            "put": {
                "description": "管理员更新一个应用配置。女菩萨的可见范围固定为所有人。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "更新应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "应用配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.adminAppRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.adminAppEnvelope"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员删除普通应用。女菩萨不能删除。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "删除应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.successEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/admin/apps/{id}/disable": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "禁用应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.adminAppEnvelope"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/admin/apps/{id}/enable": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "启用应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.adminAppEnvelope"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/admin/apps/{id}/secret/regenerate": {
            "post": {
                "description": "普通应用可以生成新密钥。女菩萨密钥由配置管理，不能在后台生成。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员应用"
                ],
                "summary": "生成应用回调密钥",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.adminAppEnvelope"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpserver.adminAppEnvelope": {
            "type": "object",
            "properties": {
                "app": {
                    "$ref": "#/definitions/httpserver.adminAppResponse"
                }
            }
        },
        "httpserver.adminAppRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "callback_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "httpserver.adminAppResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "callback_secret": {
                    "type": "string"
                },
                "callback_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "creator_user_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "system": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "httpserver.errorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpserver.errorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpserver.errorBody"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "httpserver.listAdminAppsResponse": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpserver.adminAppResponse"
                    }
                }
            }
        },
...
        "httpserver.successEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}